Fix introduced bugs from Issue 3
There are 2 main categories of bugs fixed by this commit. The first is names which have a templated namespace - these no-longer perform extraneous deletion of previous names. The second is only deleting namespace-names when a :: is found in a name of a special category, and not any time an operator is found within a name